services/api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
datasets==2.18.0
dill==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-jwt==1.19.2
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gunicorn==23.0.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.4
joblib==1.4.0
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.1.0
pooch==1.8.2
psycopg2==2.9.9
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3.post1
pyvista==0.44.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.13.0
scooby==0.10.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.13.3
torch==2.3.0
tqdm==4.66.2
transformers==4.31.0
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vtk==9.3.1
Werkzeug==3.0.4
xxhash==3.4.1
yarl==1.9.4
